feat: Adicionando validation

+++ SmartOficina.Api/Controllers/ClienteController.cs
@@ -16,30 +16,59 @@ public class ClienteController : ControllerBase
     [HttpPost]
     public async Task<IActionResult> AddAsync(ClienteDto cliente)
     {
+        if (!ModelState.IsValid)
+        {
+            return StatusCode(StatusCodes.Status400BadRequest, ModelState);
+        }
         return Ok(await _repository.Create(_mapper.Map<Cliente>(cliente)));
     }
 
     [HttpGet]
     public async Task<IActionResult> GetAll()
     {
+        if (!ModelState.IsValid)
+        {
+            return StatusCode(StatusCodes.Status400BadRequest, ModelState);
+        }
         return Ok(await _repository.GetAll());
     }
 
     [HttpGet("id")]
     public async Task<IActionResult> GetId(Guid id)
     {
+        if (!ModelState.IsValid || id == null)
+        {
+            if (ModelState.ErrorCount < 1)
+                ModelState.AddModelError("error", "Id invalid");
+
+            return StatusCode(StatusCodes.Status400BadRequest, ModelState);
+        }
         return Ok(await _repository.FindById(id));
     }
 
     [HttpPut]
     public async Task<IActionResult> AtualizarCliente(ClienteDto cliente)
     {
+        if (!ModelState.IsValid || !cliente.Id.HasValue)
+        {
+            if (ModelState.ErrorCount < 1)
+                ModelState.AddModelError("error", "Id invalid");
+
+            return StatusCode(StatusCodes.Status400BadRequest, ModelState);
+        }
         return Ok(await _repository.Update(_mapper.Map<Cliente>(cliente)));
     }
 
     [HttpPut("DesativarCliente")]
     public async Task<IActionResult> DesativarCliente(Guid id)
     {
+        if (!ModelState.IsValid || id == null)
+        {
+            if (ModelState.ErrorCount < 1)
+                ModelState.AddModelError("error", "Id invalid");
+
+            return StatusCode(StatusCodes.Status400BadRequest, ModelState);
+        }
         return Ok(await _repository.Desabled(id));
     }
 
@@ -48,6 +77,13 @@ public class ClienteController : ControllerBase
     {
         try
         {
+            if (!ModelState.IsValid || id == null)
+            {
+                if (ModelState.ErrorCount < 1)
+                    ModelState.AddModelError("error", "Id invalid");
+
+                return StatusCode(StatusCodes.Status400BadRequest, ModelState);
+            }
             await _repository.Delete(id);
             return Ok();
         }
