+++ SmartOficina.Api/Controllers/SubServicoController.cs
@@ -15,6 +15,10 @@ public class SubServicoController : ControllerBase
     [HttpPost]
     public async Task<IActionResult> AddAsync(SubServicoDto subServico)
     {
+        if (!ModelState.IsValid)
+        {
+            return StatusCode(StatusCodes.Status400BadRequest, ModelState);
+        }
         return Ok(await _repository.Create(_mapper.Map<SubServico>(subServico)));
     }
 
@@ -27,18 +31,39 @@ public class SubServicoController : ControllerBase
     [HttpGet("id")]
     public async Task<IActionResult> GetId(Guid id)
     {
+        if (!ModelState.IsValid || id == null)
+        {
+            if (ModelState.ErrorCount < 1)
+                ModelState.AddModelError("error", "Id invalid");
+
+            return StatusCode(StatusCodes.Status400BadRequest, ModelState);
+        }
         return Ok(await _repository.FindById(id));
     }
 
     [HttpPut]
     public async Task<IActionResult> AtualizarSubServico(SubServicoDto subServico)
     {
+        if (!ModelState.IsValid || !subServico.Id.HasValue)
+        {
+            if (ModelState.ErrorCount < 1)
+                ModelState.AddModelError("error", "Id invalid");
+
+            return StatusCode(StatusCodes.Status400BadRequest, ModelState);
+        }
         return Ok(await _repository.Update(_mapper.Map<SubServico>(subServico)));
     }
 
     [HttpPut("DesativarSubServico")]
     public async Task<IActionResult> DesativarSubServico(Guid id)
     {
+        if (!ModelState.IsValid || id == null)
+        {
+            if (ModelState.ErrorCount < 1)
+                ModelState.AddModelError("error", "Id invalid");
+
+            return StatusCode(StatusCodes.Status400BadRequest, ModelState);
+        }
         return Ok(await _repository.Desabled(id));
     }
 
@@ -47,6 +72,13 @@ public class SubServicoController : ControllerBase
     {
         try
         {
+            if (!ModelState.IsValid || id == null)
+            {
+                if (ModelState.ErrorCount < 1)
+                    ModelState.AddModelError("error", "Id invalid");
+
+                return StatusCode(StatusCodes.Status400BadRequest, ModelState);
+            }
             await _repository.Delete(id);
             return Ok();
         }
